requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
numpy==1.24.4
onnxruntime==1.19.2
pillow==10.4.0